+++ app/page.tsx
@@ -1,6 +1,6 @@
 import { Typography } from "@/components/base";
 import Link from "next/link";
-import Hero from "@/components/home/Hero/Hero";
+import Hero from "@/components/hero/Hero";
 import { ProjectCard } from "@/components/cards/ProjectCard";
 import {
   BrowseGrid,
