+++ components/hero/data.ts
